CCIP-8338 Processing attestation for CCTP (#448)
* Main attestation logic added

* Main attestation logic added

* Main attestation logic added

* Main attestation logic added

* Main attestation logic added

* Fixes

* Fixes

* Fixes

* Fixes

* Tests

* Removing unused ccvAddress from Attestation

* Post review fixes

+++ verifier/token/cctp/http_client_test.go
@@ -405,7 +405,7 @@ func TestParseResponseBody(t *testing.T) {
 		assert.Equal(t, "0xabcdef1234567890", msg.Message)
 		assert.Equal(t, "12345", msg.EventNonce)
 		assert.Equal(t, "0x9876543210fedcba", msg.Attestation)
-		assert.Equal(t, 2, msg.CCTPVersion)
+		assert.Equal(t, float64(2), msg.CCTPVersion)
 		assert.Equal(t, "complete", msg.Status)
 		assert.Equal(t, "0", msg.DecodedMessage.SourceDomain)
 		assert.Equal(t, "1000000", msg.DecodedMessage.DecodedMessageBody.Amount)
